重构后台脚本
